requirements.txt
azure-identity>=1.12.0
azure-mgmt-resource>=23.0.0
azure-mgmt-resourcegraph>=3.0.0
azure-mgmt-monitor>=2.0.0
pandas>=2.0
tqdm
